+++ examples/basic/.firebaserc
@@ -1,0 +1,32 @@
+{
+  "projects": {
+    "default": "redux-firebasev3",
+    "master": "redux-firebasev3",
+    "prod": "redux-firebasev3"
+  },
+  "ci": {
+    "copyVersion": true,
+    "createConfig": {
+      "master": {
+        "env": "staging",
+        "firebase": {
+          "apiKey": "AIzaSyCTUERDM-Pchn_UDTsfhVPiwM4TtNIxots",
+          "authDomain": "redux-firebasev3.firebaseapp.com",
+          "databaseURL": "https://redux-firebasev3.firebaseio.com",
+          "projectId": "redux-firebasev3",
+          "storageBucket": "redux-firebasev3.appspot.com"
+        }
+      },
+      "prod": {
+        "env": "production",
+        "firebase": {
+          "apiKey": "AIzaSyCTUERDM-Pchn_UDTsfhVPiwM4TtNIxots",
+          "authDomain": "redux-firebasev3.firebaseapp.com",
+          "databaseURL": "https://redux-firebasev3.firebaseio.com",
+          "projectId": "redux-firebasev3",
+          "storageBucket": "redux-firebasev3.appspot.com"
+        }
+      }
+    }
+  }
+}
