+++ examples/typescript/cypress/plugins/index.js
@@ -1,8 +1,20 @@
-const wp = require('@cypress/webpack-preprocessor')
+const fs = require('fs')
+const cypressTypeScriptPreprocessor = require('./cy-ts-preprocessor')
 
-module.exports = (on) => {
-  const options = {
-    webpackOptions: require('../../webpack.config'),
-  }
-  on('file:preprocessor', wp(options))
+module.exports = on => {
+  on('file:preprocessor', (file) => {
+    console.log(`preprocessor invoked with ${JSON.stringify(file)}`);
+    return cypressTypeScriptPreprocessor(file).then((results) => {
+      console.log(`preprocessor returned ${JSON.stringify(results)}`);
+
+      if (!fs.existsSync(file.outputPath)) {
+        console.error(`Output file does not exist on the filesystem: ${JSON.stringify(file)}`);
+        throw new Error(`Output file does not exist on the filesystem: ${JSON.stringify(file)}`);
+      }
+      return results;
+    }).catch((err) => {
+      console.error(`Error occurred running preprocessor`, err);
+      throw err;
+    });
+  })
 }
