--- examples/typescript/webpack.config.js
@@ -1,19 +1,0 @@
-module.exports = {
-  resolve: {
-    extensions: ['.ts', '.js'],
-  },
-  module: {
-    rules: [
-      {
-        test: /\.ts$/,
-        exclude: [/node_modules/],
-        use: [{
-          loader: 'ts-loader',
-          options: {
-            transpileOnly: true,
-          },
-        }],
-      },
-    ],
-  },
-}
