+++ package.json
@@ -29,7 +29,7 @@
     "chalk": "^2.4.2",
     "commander": "^3.0.0",
     "figures": "^3.0.0",
-    "firebase-admin": "^8.4.0",
+    "firebase-admin": "^8.5.0",
     "firebase-tools-extra": "0.1.1",
     "lodash": "^4.17.15"
   },
@@ -39,15 +39,14 @@
     "@typescript-eslint/eslint-plugin": "^2.0.0",
     "@typescript-eslint/parser": "^2.0.0",
     "cross-env": "^5.2.0",
-    "eslint": "^6.1.0",
-    "eslint-config-airbnb": "^18.0.1",
+    "eslint": "^6.3.0",
     "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-config-prettier": "^6.1.0",
+    "eslint-config-prettier": "^6.2.0",
     "eslint-plugin-babel": "^5.3.0",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^3.1.0",
-    "firebase-tools": "^7.2.4",
+    "firebase-tools": "^7.3.1",
     "husky": "^3.0.4",
     "prettier": "^1.18.2",
     "rimraf": "^3.0.0",
