+++ src/constants.ts
@@ -8,4 +8,3 @@ export const FIREBASE_TOOLS_BASE_COMMAND = '$(npm bin)/firebase';
 // Path to firebase-tools-extra Command Line tool (wrapper for firebase-tools)
 export const FIREBASE_EXTRA_PATH = '$(npm bin)/firebase-extra';
 export const FIREBASE_TOOLS_YES_ARGUMENT = '-y';
-export const DEFAULT_BASE_PATH = process.cwd();
